requirements.txt
# Requirements for production / heroku

# Django
Django==1.7.1

# Better templates
MarkupSafe==0.23
Jinja2==2.7.3
jingo==0.7

# Django REST Framework
djangorestframework==2.3.14
Markdown==2.5.1
django-filter==0.8

# Django extensions
six==1.8.0
django-extensions==1.4.6

# JSON API interfaces for Django REST Framework
https://github.com/jwhitlock/drf-json-api/archive/v0.1d.zip

# History of changes to models
# django-simple-history==1.5.0 has deprecation warnings w/ Django 1.7
# PR #141 - Fix issue with Meta.order_with_respect_to
# Use my django-simple-history w/ PR #141 on 2014-11-16
git+git://github.com/jwhitlock/django-simple-history.git@order_wrt_included#egg=django-simple-history

# Configure database from env
dj-database-url==0.3.0

# Better test runner (included in settings)
nose==1.3.4
# django-nose==1.2 has deprecation warnings w/ Django 1.7
# Use django-nose head on 2014-11-15
git+git://github.com/django-nose/django-nose.git@b81f8471e#egg=django-nose

# Test Mocking, included in Python 3.3
mock==1.0.1

# WSGI runner, used in Heroku
gunicorn==19.1.1

# Connect to PostgreSQL
psycopg2==2.5.4

# Serve static files
static3==0.5.1
dj-static==0.0.6

# Timezone info
pytz==2014.9

# Fast Memcache connections, and tell Heroku to install C dependencies
# pylibmc==1.3.0 does not support Python 3.3 or 3.4
# Use pylibmc head on 2014-11-15
git+git://github.com/lericson/pylibmc.git@c6da0ed9c#egg=pylibmc
# django-pylibmc==0.5.0 has Python 3.x and Django 1.6/1.7 problems
# Use my py3 branch on 2014-11-15
git+git://github.com/jwhitlock/django-pylibmc@a6020ce#egg=django-pylibmc
django-pylibmc-sasl==0.2.4

# Celery - async task management
billiard==3.3.0.18
amqp==1.4.6
anyjson==0.3.3
kombu==3.0.23
celery==3.1.16

# Modified Preorder Tree Traversal
django-mptt==0.6.1

# Sorted ManyToManyField
django-sortedm2m==0.8.1

# Cached instances for Django REST Framework
drf-cached-instances==0.2.0

# Better HTTP Client for tools
requests==2.4.3

# CORS headers in middleware
git+git://github.com/ottoyiu/django-cors-headers@5d5eeccf7ce#egg=django-cors-headers

# Parsing Expression Grammar, for MDN scraping
parsimonious==0.6.2
